Close #69 ,close #70 ,close #71 ,close #72 ,close #73 ,close #74 ,close #75 ,close #76

+++ core/src/test/java/com/holonplatform/datastore/jdbc/test/suite/database/MariaDBTest.java
@@ -18,6 +18,8 @@ package com.holonplatform.datastore.jdbc.test.suite.database;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
+import java.math.BigInteger;
+
 import org.junit.Test;
 
 import com.holonplatform.core.datastore.DataTarget;
@@ -53,7 +55,9 @@ public class MariaDBTest extends AbstractDatabaseSuiteTest {
 
 				assertEquals(1, result.getInsertedKeys().size());
 
-				assertEquals(Long.valueOf(1), result.getInsertedKeys().values().iterator().next());
+				// for MariaDB10
+				assertEquals(BigInteger.valueOf(1), result.getInsertedKeys().values().iterator().next());
+				// for MariaDB assertEquals(Long.valueOf(1), result.getInsertedKeys().values().iterator().next());
 				assertEquals("code", result.getInsertedKeys().keySet().iterator().next().getName());
 
 				// bring back ids
